+++ src/win32/RP_ExtractImage.cpp
@@ -224,11 +224,17 @@ IFACEMETHODIMP RP_ExtractImage::Extract(HBITMAP *phBmpImage)
 
 	// ROM is supported. Get the image.
 	// NOTE: Using width only. (TODO: both width/height?)
-	int ret = d->thumbnailer.getThumbnail(d->romData, d->rgSize.cx, *phBmpImage);
-	if (ret != 0 || !*phBmpImage) {
+	CreateThumbnail::GetThumbnailOutParams_t outParams;
+	outParams.retImg = nullptr;
+	int ret = d->thumbnailer.getThumbnail(d->romData, d->rgSize.cx, &outParams);
+	if (ret != 0 || !outParams.retImg) {
 		// ROM is not supported. Use the fallback.
+		if (outParams.retImg) {
+			DeleteBitmap(outParams.retImg);
+		}
 		return d->Fallback(phBmpImage);
 	}
+	*phBmpImage = outParams.retImg;
 	return S_OK;
 }